+++ src/main/java/com/tencao/projectbalance/gameObjs/tile/CondenserTile.kt
@@ -43,11 +43,11 @@ open class CondenserTile : TileEmc(), IEmcAcceptor, ICraftingGen {
     val lock: ItemStackHandler = this.StackHandler(1)
     private var isAcceptingEmc: Boolean = false
     private var ticksSinceSync: Int = 0
-    var displayEmc: Int = 0
+    var displayEmc: Long = 0
     var lidAngle: Float = 0.toFloat()
     var prevLidAngle: Float = 0.toFloat()
     var numPlayersUsing: Int = 0
-    var requiredEmc: Int = 0
+    var requiredEmc: Long = 0
     var requiredTime: Long = 0
     var timePassed: Long = 0
     var craftTimer: Int = 0
@@ -99,9 +99,9 @@ open class CondenserTile : TileEmc(), IEmcAcceptor, ICraftingGen {
 
         checkLockAndUpdate()
 
-        displayEmc = this.storedEmc.toInt()
+        displayEmc = this.storedEmc.toLong()
 
-        if (!lock.getStackInSlot(0).isEmpty && requiredEmc != 0) {
+        if (!lock.getStackInSlot(0).isEmpty && requiredEmc != 0L) {
             condense()
         }
     }
